Fix Add Products button not being active
The button would only be active based if the last TreeViewListItem that
is visible on the page is checked. It would disregard all other
checkboxes as the state of the would be recalculated and emptied for
every Item. To prevent that it now only changes the state if the user
clicks a checkbox empty or not.